review(helper): addresses review and helper functions
